feat: update route user/[id] to /[nickname]
